refactor: nomalize로 함수명 변경
