+++ packages/core/src/inputs/form-parameters.ts
@@ -1,8 +1,8 @@
 import { type FormParamDescriptor, ParamKinds } from "@/descriptors";
-import { type ArrayOrRecord, applyParameters } from "@/utils/parameter-normalizer";
+import { type ArrayOrRecord, applyNormalize } from "@/utils/normalizer";
 
 type FormInput = ArrayOrRecord<typeof ParamKinds.Form, FormParamDescriptor>;
 
 export const applyFormParameters = (input: FormInput): FormParamDescriptor[] => {
-    return applyParameters<typeof ParamKinds.Form, FormParamDescriptor>(ParamKinds.Form, input);
+    return applyNormalize<typeof ParamKinds.Form, FormParamDescriptor>(ParamKinds.Form, input);
 };
