+++ packages/core/src/inputs/path-parameters.ts
@@ -1,8 +1,8 @@
 import { ParamKinds, type PathParamDescriptor } from "@/descriptors";
-import { type ArrayOrRecord, applyParameters } from "@/utils/parameter-normalizer";
+import { type ArrayOrRecord, applyNormalize } from "@/utils/normalizer";
 
 type PathInput = ArrayOrRecord<typeof ParamKinds.Path, PathParamDescriptor>;
 
 export const applyPathParameters = (input: PathInput): PathParamDescriptor[] => {
-    return applyParameters<typeof ParamKinds.Path, PathParamDescriptor>(ParamKinds.Path, input);
+    return applyNormalize<typeof ParamKinds.Path, PathParamDescriptor>(ParamKinds.Path, input);
 };
